last-upd

+++ server/app/Models/userModel.js
@@ -7,23 +7,29 @@ class User {
     nickName;
     heartTimerId;
     respondTimerId;
+    inline;
     constructor(socket, nickName) {
         let that = this;
         that.id = tmpNum++;
         that.socket = socket;
         that.nickName = nickName;
         that.respondTimerId = null;
+        that.inline = true;
     }
     //考虑到后面可能将数据存在缓存数据库中，因此设置成可以异步的方法。
     save(callback) {
         let that = this;
-        // console.log(typeof that.nickName + "  " +typeof that.socket);
+        that.inline = true;
         if (typeof that.nickName != "string" || typeof that.socket != "object")
             return false;
         return that.repository.save(this, callback);
     }
     delete(callback) {
-        return this.repository.delete(this, callback);
+        let that = this;
+        if (that.inline) {
+            that.inline = false;
+            return this.repository.delete(this, callback);
+        } else return false;
     }
 }
 exports.form = function (socket, nickName) {
